+++ CHANGELOG.md
@@ -4,8 +4,33 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
-- Display error message in case of XML parsing error
 - New command [hardis:doc:fieldusage](https://sfdx-hardis.cloudity.com/hardis/doc/fieldusage/) to generate a report with custom field's usage from metadata dependencies.
+
+## [5.17.3] 2025-01-29
+
+- [hardis:doc:project2markdown](https://sfdx-hardis.cloudity.com/hardis/doc/project2markdown/): Improve Apex docs markdown
+- Upgrade apexdocs version
+- Fix auth message when selecting default org
+
+## [5.17.2] 2025-01-29
+
+- [hardis:org:configure:files](https://sfdx-hardis.cloudity.com/hardis/org/configure/files/): Add examples when configuring file export format
+- [hardis:doc:project2markdown](https://sfdx-hardis.cloudity.com/hardis/doc/project2markdown/): Avoid the command to crash if apexdocs generation fails
+
+## [5.17.1] 2025-01-27
+
+- [hardis:doc:project2markdown](https://sfdx-hardis.cloudity.com/hardis/doc/project2markdown/): Add type of Lightning Pages in tables
+- [hardis:org:monitor:backup](https://sfdx-hardis.cloudity.com/hardis/org/monitor/backup/): Fix issue when there is an empty metadata type
+
+## [5.17.0] 2025-01-26
+
+- [hardis:doc:project2markdown](https://sfdx-hardis.cloudity.com/hardis/doc/project2markdown/) enhancements:
+  - Generate Apex classes documentation using `@cparra/apexdocs`, and describe them using AI if available
+  - Generate Lightning Pages documentation and describe them using AI if available
+  - Display error message in case of XML parsing error
+  - Do not raise issues when managed items fields don't have descriptions
+  - Do not raise inactive validation rule issue when the VR is from a managed package
+  - Fix New JSON coverage formatter is selecting wrong JSON from sf project deploy command
 
 ## [5.16.4] 2025-01-22